requirements.txt
Flask==2.0.2
jsonify==0.5
mysql-connector-python==8.0.27